requirements.txt
python-dotenv==1.2.1
redis==5.0.1
chainlit>=1.1.0
requests
beautifulsoup4
pytest
aiomysql

# LangChain pour orchestration d'agent
langchain>=1.0.0
langchain-google-genai>=4.2.0
langchain-openai>=0.1.0
langchain-core>=1.0.0

# LLM APIs
google-generativeai>=0.8.0
openai>=1.0.0

# RAG vectoriel - FAISS
sentence-transformers>=5.0.0
# faiss-cpu>=1.12.0

# Observabilité - Langfuse
langfuse>=3.0.0
